requirements.txt
pandas 
mysql.connector 
streamlit 
plotly.express
os
json
plotly.graph_objects